requirements.txt
langchain
langchain-community
python-dotenv
requests
langchain-openai
langchain-anthropic
langchain-google-genai